refactor and add tests

+++ src/main/java/io/github/arlol/chorito/chores/GraalNativeImagePropertiesChore.java
@@ -1,5 +1,7 @@
 package io.github.arlol.chorito.chores;
 
+import java.nio.file.Path;
+
 import io.github.arlol.chorito.tools.ChoreContext;
 import io.github.arlol.chorito.tools.FilesSilent;
 
@@ -12,14 +14,19 @@ public class GraalNativeImagePropertiesChore {
 	}
 
 	public void doit() {
-		FilesSilent.walk(
-				context.resolve("src/main/resources/META-INF/native-image/")
-		).filter(p -> p.endsWith("native-image.properties")).forEach(p -> {
-			String content = FilesSilent.readString(p);
-			content = content.replace("--allow-incomplete-classpath", "");
-			content = content.replace("\n \\\n", "\n");
-			FilesSilent.writeString(p, content);
-		});
+		Path metaInfNativeImage = context
+				.resolve("src/main/resources/META-INF/native-image/");
+		if (FilesSilent.exists(metaInfNativeImage)) {
+			FilesSilent.walk(metaInfNativeImage)
+					.filter(p -> p.endsWith("native-image.properties"))
+					.forEach(p -> {
+						String content = FilesSilent.readString(p);
+						content = content
+								.replace("--allow-incomplete-classpath", "");
+						content = content.replace("\n \\\n", "\n");
+						FilesSilent.writeString(p, content);
+					});
+		}
 	}
 
 }
